Add module overrides for fips
We need to use the SHA1 module under OpenSSL because the openssl
functions called by Digest::SHA1 cause openssl to crash the process.

We use the Digest::MD5 over the OpenSSL::MD5 module because md5
is not allowed when in fips mode and causes the process to crash.
While we work through these issues, we're going to allow it to
pass by compiling the ruby md5 implementation.